# schema_service/openapi.yaml
openapi: 3.0.3
info:
  title: Schema Service HTTP API
  version: 0.1.0
  description: |
    Canonical schema service for FoldDB — the single source of truth for
    schema creation, similarity matching, and the schema registry.

    This spec documents the `/v1/*` HTTP surface exposed by:

      * the dev binary `schema_service` (crate `schema_service_server_http`)
      * the production Lambda at `schema.folddb.com`
        (crate `schema_service_server_lambda`)

    Both mount the same handler set from `schema_service_server_shared`, so
    the wire contract is identical. Third-party integrators can generate a
    client with any OpenAPI 3 codegen, e.g.:

    ```
    openapi-generator-cli generate -i openapi.yaml -g rust -o ./client-rust
    openapi-generator-cli generate -i openapi.yaml -g typescript-fetch -o ./client-ts
    ```

    A reference Rust client lives alongside this spec in
    `crates/client/` (published as `schema_service_client`).

    Some nested types (Schema, Query, FieldValueType, DataClassification,
    DeclarativeSchemaType, KeyConfig) are owned by the
    `fold_db` crate. They appear here as loosely-typed objects; the
    authoritative definitions live in
    `https://github.com/EdgeVector/fold/tree/main/fold_db/src/schema/types`.

    ## Immutability contract

    **Everything this service approves is immutable forever.** Schemas,
    classifications, and any other artifact returned by `/v1/*` is
...

...

...
        reissues bytes under an existing hash.
...

...

...
security: []
servers:
...
paths:
...
  /v1/schemas:
    get:
      tags: [schemas]
      operationId: list_schemas
      summary: List schema names
      responses:
        '200':
          description: List of registered schema names.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SchemasListResponse'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags: [schemas]
      operationId: add_schema
      summary: Register a schema (idempotent; may expand an existing one)
      description: |
        Registers a schema along with mutation mappers. The service canonicalizes
        the schema and may either create it, return the existing identity, or
        expand an existing schema by superseding it.

        Status codes:

        * `201 Created` — newly added, or expanded from an existing schema
          (check `replaced_schema` to see if this was an expansion).
        * `200 OK` — schema with the same identity hash already existed.
        * `400 Bad Request` — validation / canonicalization failure.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddSchemaRequest'
      responses:
        '200':
          description: Schema with this identity hash already existed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddSchemaResponse'
        '201':
          description: Schema was newly added (or expanded from an existing one, see `replaced_schema`).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddSchemaResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /v1/schemas/mutation-challenge:
    post:
      tags: [schemas]
      operationId: issue_schema_mutation_challenge
      summary: Issue a node-key proof-of-work challenge
      description: |
        Issues a short-lived challenge for feature-flagged enforcement of
        unowned shared schema mutations. The proof is bound to the node
        publish key, canonical schema JSON hash, nonce, and counter.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SchemaMutationChallengeRequest'
      responses:
        '201':
          description: Challenge issued.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SchemaMutationChallengeResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /v1/registry/index:
    get:
      tags: [registry]
      operationId: registry_index
      summary: Signed compact registry index
      description: |
        Returns a deterministic, signed, compact projection of the active
        schema registry for local dedup caches. The endpoint is read-only:
        repeated calls against the same registry state return the same body
        and do not advance the registry version.

        The response includes active schema identity hashes, descriptive
        names, owner app namespaces, field sets, field descriptions,
        supersession redirects, a monotonic `registry_version`, a BLAKE3
        Merkle root over the canonical entry set, and an Ed25519 signature
        over the canonical JSON body with `signature` removed.
      parameters:
        - name: If-None-Match
          in: header
          required: false
          schema:
            type: string
          description: Quoted `registry_version` ETag. Returns `304` when current.
      responses:
        '200':
          description: Signed compact registry index.
          headers:
            Cache-Control:
              schema:
                type: string
              description: '`public, max-age=60`'
            ETag:
              schema:
                type: string
              description: Quoted registry version.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegistryIndexEnvelope'
        '304':
          description: The caller's ETag already matches the current registry version.
        '500':
          $ref: '#/components/responses/InternalError'
  /v1/schemas/batch-check-reuse:
    post:
      tags: [schemas]
      operationId: batch_check_reuse
      summary: Check many proposed schemas against the registry in one call
      description: |
        For each `SchemaLookupEntry` in the request, returns a match if one
        exists in the registry. Names with no match are omitted from the
        response `matches` map. Intended for AI ingestion pipelines that
        propose many candidate schemas at once.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchSchemaReuseRequest'
      responses:
        '200':
          description: Matches for the proposed schemas (only names with matches are included).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchSchemaReuseResponse'
        '500':
          $ref: '#/components/responses/InternalError'
  /v1/schemas/resolve:
    post:
      tags: [schemas]
      operationId: resolve_schemas
      summary: Stateless schema dedup for cache misses
      description: |
        Resolves one or more schema proposals against the current shared
        registry without mutating shared state. This endpoint is a public
        read path for local cache misses: it may return reuse advice,
        candidate equivalents, a novel decision, or a refresh decision when
        the caller's `client_registry_version` is stale.

        The handler never creates schemas, canonical fields, app records, or
        shared discovery state. It also does not advance `registry_version`;
        repeated calls against identical registry state return identical
        decisions.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SchemaResolveRequest'
      responses:
        '200':
          description: Per-proposal resolve decisions.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SchemaResolveResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /v1/schemas/reload:
    post:
      tags: [schemas]
      operationId: reload_schemas
      summary: Reload schemas from persistent storage
      description: Drops the in-memory schema cache and repopulates it from the backing store.
      responses:
        '200':
          description: Reload completed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReloadResponse'
        '500':
          $ref: '#/components/responses/InternalError'
  /v1/schemas/available:
    get:
      tags: [schemas]
      operationId: get_available_schemas
      summary: Full schema definitions (not just names)
      description: |
        Returns every schema with its full declarative definition, wrapped in
        a `SchemaEnvelope` that also carries a legacy `system` flag.

        Each envelope's `schema.source` field is one of `system_seed`,
        `starter_seed`, or `user` (see `SchemaSource` in the fold_db schema
        types). Clients can filter at the wire via the `source` query
        parameter below; unknown values return an empty list.
      parameters:
        - name: source
          in: query
          required: false
          description: |
            Optional filter by schema origin. Accepted values:
            `system_seed` (service primitives), `starter_seed` (pre-loaded
            starter buckets such as Schema.org types), or `user`
            (node-created). Any other value returns an empty list; omit
            the parameter to return all schemas.
          schema:
            type: string
            enum: [system_seed, starter_seed, user]
      responses:
        '200':
          description: All schemas (optionally filtered by source) with full definitions.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AvailableSchemasResponse'
        '500':
          $ref: '#/components/responses/InternalError'
  /v1/schemas/similar/{name}:
    get:
      tags: [schemas]
      operationId: find_similar_schemas
      summary: Find schemas similar to the named schema
      description: |
        Returns schemas whose field-level embeddings cluster close to the
        named schema, above the given cosine-similarity threshold.
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
          description: The reference schema name.
        - name: threshold
          in: query
          required: false
          schema:
            type: number
            format: double
            minimum: 0.0
            maximum: 1.0
            default: 0.5
          description: Cosine-similarity threshold (0.0 – 1.0). Defaults to 0.5.
      responses:
        '200':
          description: Matching similar schemas.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimilarSchemasResponse'
        '400':
          description: Threshold outside `[0.0, 1.0]`.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Reference schema not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalError'
  /v1/schema/{name}:
    get:
      tags: [schemas]
      operationId: get_schema
      summary: Get one schema by name
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: |
            The full schema definition plus a `system: bool` flag
            indicating whether this is an infrastructure schema
            (`true`) or a user-data schema (`false`). The schema's
            own fields are flattened at the top level of the
            response; see `SchemaEnvelope`.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SchemaEnvelope'
        '404':
          description: Schema not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /v1/snapshot:
    get:
      tags: [snapshot]
      operationId: snapshot_export
      summary: Export the registry as a single self-contained JSON snapshot
      description: |
        Returns the full registry — schemas, apps, canonical fields,
        and the embedding caches — in one JSON envelope.

        Used by the dev binary's `--hydrate-from` flow (see
        `projects/schema-service-dev-hydration`) to seed a local Sled
        instance from production. Auth is enforced by the framework
        wrapper: production gates this route on `X-API-Key`; the dev
        binary leaves it open for localhost loopback.

        For **resolver-pack publishing**, prefer
        `GET /v1/snapshot/shared-only` so private legacy bootstrap
        schemas never enter pack artifacts.
      responses:
        '200':
          description: A snapshot envelope.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SnapshotEnvelope'
        '500':
          $ref: '#/components/responses/InternalError'
  /v1/snapshot/shared-only:
    get:
      tags: [snapshot]
      operationId: snapshot_export_shared_only
      summary: Export the shared-only registry projection for resolver packs
      description: |
        Same envelope shape as `GET /v1/snapshot`, but schemas are
        filtered to the shared-surface projection:

        - includes system-owned seeds and explicit shared contracts
        - excludes private legacy bootstrap, unknown, and superseded rows

        Embeddings are stripped on the wire (same as full snapshot).
        Auth matches `GET /v1/snapshot` (X-API-Key in production).

        Pack publishers also re-apply the projection client-side when
        building from a full snapshot file, so either input is safe.
      responses:
        '200':
          description: Shared-only snapshot envelope.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SnapshotEnvelope'
        '500':
          $ref: '#/components/responses/InternalError'
  /v1/snapshot/import:
    post:
      tags: [snapshot]
      operationId: snapshot_import
      summary: Replace the local registry with the contents of a snapshot envelope
      description: |
        Full-replace import. Clears every owned Sled tree (`schemas`,
        `canonical_fields`, `apps`), writes the envelope's contents
        in their place, refreshes the in-memory caches, and rebuilds the
        `descriptive_name` index.

        **Sled-only.** The production Lambda binary intentionally does
        not mount this route; calling it through an external backend
        returns `400`.

        The importer rejects envelopes whose `format_version` does not
        equal the binary's [`SNAPSHOT_FORMAT_VERSION`] or whose
        `embedder_version` differs from the local embedder — vectors
        from different models occupy different spaces and cannot be
        mixed.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SnapshotEnvelope'
      responses:
        '200':
          description: Import completed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SnapshotImportReport'
        '400':
          description: |
            Envelope rejected — unknown `format_version`, embedder
            mismatch, non-Sled storage backend, or malformed body.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/system/reset:
    post:
      tags: [system]
      operationId: reset_database
      summary: Reset the local schema database (local storage backends only)
      description: |
        Clears all schemas. **Only supported on Sled-backed dev instances**;
        the production Lambda returns `400 Bad Request` because its external
        backend owns its own reset semantics.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetRequest'
      responses:
        '200':
          description: Reset completed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResetResponse'
        '400':
          description: Missing confirmation, or endpoint not supported for this backend.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResetResponse'
        '500':
          description: Storage backend failure.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResetResponse'
  /v1/admin/dedupe-descriptive-names:
    post:
      tags: [system]
      operationId: dedupe_descriptive_names
      summary: Mark duplicate active descriptive-name entries superseded
      description: |
        One-shot cleanup for registry states where multiple active schemas
        share a descriptive name. Keeps the active schema with the largest
        field set and marks the others `superseded_by` that survivor.
      responses:
        '200':
          description: Cleanup report.
          content:
            application/json:
              schema:
                type: object
                required: [groups]
                properties:
                  groups:
                    type: array
                    items:
                      type: object
                      required: [descriptive_name, survivor, survivor_field_count, superseded]
                      properties:
                        descriptive_name:
                          type: string
                        survivor:
                          type: string
                        survivor_field_count:
                          type: integer
                          format: int64
                        superseded:
                          type: array
                          items:
                            type: string
        '500':
          $ref: '#/components/responses/InternalError'
  /v1/admin/deprecate-schemas:
    post:
      tags: [system]
      operationId: deprecate_schemas
      summary: Mark selected schemas inactive without deleting records
      description: |
        Surgical cleanup for polluted registry entries. Accepts schema
        identity hashes and/or descriptive names, marks matching active schemas
        inactive using `superseded_by`, and removes matching descriptive-name
        index entries. The immutable schema records remain readable by hash.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                schema_names:
                  type: array
                  items:
                    type: string
                descriptive_names:
                  type: array
                  items:
                    type: string
                owner_app_id:
                  type: string
                  nullable: true
      responses:
        '200':
          description: Deprecation report.
          content:
            application/json:
              schema:
                type: object
                required: [deprecated, not_found]
                properties:
                  deprecated:
                    type: array
                    items:
                      type: object
                      required: [schema_name, already_deprecated]
                      properties:
                        schema_name:
                          type: string
                        descriptive_name:
                          type: string
                          nullable: true
                        already_deprecated:
                          type: boolean
                  not_found:
                    type: array
                    items:
                      type: string
        '500':
          $ref: '#/components/responses/InternalError'
...
  /v1/fields/{declaration_id}:
    get:
      tags: [fields]
      operationId: get_declared_field
      summary: Read one field declaration
      description: |
        Unauthenticated, like `GET /v1/apps/{app_id}`: a node needs to resolve a
        declaration it already holds an id for, and the id is not a secret. What
        is gated is *declaring* and *referencing*, never looking up.
      parameters:
        - name: declaration_id
          in: path
          required: true
          schema:
            type: string
          description: The minted declaration id, e.g. `fd_2f8a1c9e...`.
      responses:
        '200':
          description: The declaration.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeclaredFieldRecord'
        '404':
          description: No declaration with that id.
...
components:
  responses:
...
    InternalError:
      description: Internal server error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  securitySchemes:
    # app_identity v3.1 developer cert + dev-signed envelope. Both headers
    # are required together on cert-gated writes.
    DevCert:
      type: apiKey
      in: header
      name: X-Exemem-Dev-Cert
      description: Base64 ES256 developer cert (app_identity v3.1), issued by a trusted exemem root.
    DevSignature:
      type: apiKey
      in: header
      name: X-Signature
      description: Base64 Ed25519 SignatureEnvelope over the request body; its `purpose` pins the operation.
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    SchemasListResponse:
      type: object
      required: [schemas]
      properties:
        schemas:
          type: array
          items:
            type: string
...